# pqarrow/frostdb: accumulate results in Iterator