--- a/db_test.go
+++ b/db_test.go
@@ -271,6 +271,10 @@ func Test_DB_WithStorage(t *testing.T) {
 		dynparquet.NewSampleSchema(),
 	)
 
+	// TODO(asubiotto): These buckets are not deleted from disk, resulting in
+	// test failures if run multiple times. There seems to be an attempt below
+	// to delete them, but it doesn't have an effect. What's the magic
+	// incantation?
 	bucket, err := filesystem.NewBucket(".")
 	require.NoError(t, err)