@asubiotto 2022-08-26
I think I would still like to add a test or two (even if the test suite caught a bunch of things). I'll leave exploring improving the intermediate result cache on top of the record builder (to avoid having to recreate a builder if an optimization cannot be applied) to a future PR.