@asubiotto 2022-08-26
Cherry-picking #164 on to this PR gives us an addition 10% improvement:
```
name          old time/op    new time/op    delta
QueryTypes-8     169ms ±11%      50ms ±13%  -70.19%  (p=0.030 n=10+2)

name          old alloc/op   new alloc/op   delta
QueryTypes-8     443MB ± 0%      81MB ± 0%  -81.80%  (p=0.030 n=10+2)

name          old allocs/op  new allocs/op  delta
QueryTypes-8     1.13M ± 0%     0.45M ± 0%  -60.34%  (p=0.000 n=10+2)
```