@asubiotto 2022-09-01
Added a test for list types and refactored the list building code while I was at it. I'll let this PR sit for a bit in case there are further comments, but planning on merging this soon.